pyproject.toml
[tool.poetry]
name = "rathercurious-mastodon"
version = "0.1.0"
description = ""
authors = ["slashtechno <77907286+slashtechno@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "rathercurious-mastodon"}]

[tool.poetry.dependencies]
python = "^3.11"
mastodon-py = "^1.8.1"
python-dotenv = "^1.0.0"
trio = "^0.22.2"
loguru = "^0.7.0"
pytz = "^2023.3"
airportsdata = "^20230717"
beautifulsoup4 = "^4.12.2"
httpx = "^0.24.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
